shared.test cannot depend on vitest directly because of the apiServer intitialization details in foundry-sdk-generator
