+++ packages/foundry-sdk-generator/src/__e2e_tests__/objectSet.test.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import { apiServer } from "@osdk/shared.test";
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
 import {
   ConfidentialClientAuth,
@@ -30,7 +31,7 @@ import type {
   Employee,
   Office,
 } from "../generatedNoCheck/@test-app/osdk/ontology/objects";
-import { apiServer, fetchAllPages, fetchPage } from "./paginationHelpers";
+import { fetchAllPages, fetchPage } from "./paginationHelpers";
 import { assertOkOrThrow } from "./resultUtils";
 
 describe("Object Sets", () => {
