+++ packages/legacy-client/src/client/net/getAttachment.test.ts
@@ -17,13 +17,13 @@
 import type { ClientContext } from "@osdk/shared.net";
 import { createClientContext } from "@osdk/shared.net";
 import {
-  expectFetchToBeCalledWithGet,
   MOCK_ORIGIN,
   mockFetchResponse,
   MockOntology,
 } from "@osdk/shared.test";
 import type { MockedFunction } from "vitest";
 import { beforeEach, describe, expect, it, vi } from "vitest";
+import { expectFetchToBeCalledWithGet } from "../../util/test/expectUtils";
 import { unwrapResultOrThrow } from "../../util/test/resultUtils";
 import { getAttachment } from "./getAttachment";
