+++ packages/legacy-client/src/client/net/getFirstPoint.test.ts
@@ -18,13 +18,13 @@ import type { TimeSeriesPoint } from "@osdk/gateway/types";
 import type { ClientContext } from "@osdk/shared.net";
 import { createClientContext } from "@osdk/shared.net";
 import {
-  expectFetchToBeCalledWithGet,
   MOCK_ORIGIN,
   mockFetchResponse,
   MockOntology,
 } from "@osdk/shared.test";
 import type { MockedFunction } from "vitest";
 import { beforeEach, describe, expect, it, vi } from "vitest";
+import { expectFetchToBeCalledWithGet } from "../../util/test/expectUtils";
 import { unwrapResultOrThrow } from "../../util/test/resultUtils";
 import { Timestamp } from "../baseTypes";
 import { getFirstPoint } from "./getFirstPoint";
